Move `*_builders.py` and `model_provider.py`
They are now contained in the `megatron.training.models` module. This
way, we keep the top-level/root directory of the repository cleaner.
