+++ examples/post_training/modelopt/generate.py
@@ -21,8 +21,8 @@ from megatron.post_training.model_builder import modelopt_gpt_hybrid_builder
 from megatron.post_training.utils import report_current_memory_info, to_empty_if_meta
 from megatron.training import get_args, get_model, initialize_megatron
 from megatron.training.arguments import parse_and_validate_args
+from megatron.training.models.model_provider import model_provider
 from megatron.training.utils import print_rank_0
-from model_provider import model_provider
 
 warnings.filterwarnings('once')
