+++ examples/post_training/modelopt/offline_feature_extract.py
@@ -17,8 +17,8 @@ from megatron.post_training.arguments import add_modelopt_args
 from megatron.post_training.checkpointing import load_modelopt_checkpoint
 from megatron.post_training.model_builder import modelopt_gpt_hybrid_builder
 from megatron.training import get_args, get_model, get_tokenizer, initialize_megatron
+from megatron.training.models.model_provider import model_provider
 from megatron.training.utils import print_rank_0
-from model_provider import model_provider
 
 
 def add_extract_args(parser):
