+++ examples/post_training/modelopt/quantize.py
@@ -65,8 +65,8 @@ from megatron.post_training.utils import print_distributed_quant_summary, report
 from megatron.training import get_args, get_model, initialize_megatron
 from megatron.training.arguments import parse_and_validate_args
 from megatron.training.checkpointing import save_checkpoint
+from megatron.training.models.model_provider import model_provider
 from megatron.training.utils import print_rank_0
-from model_provider import model_provider
 
 warnings.filterwarnings("ignore")
