+++ tools/checkpoint/loader_core.py
@@ -3,12 +3,12 @@
 import json
 import os
 import sys
-import torch
 import types
 from functools import partial
 
-from schema_core import get_model_schema
+import torch
 from loader_base import MegatronCheckpointLoaderBase
+from schema_core import get_model_schema
 
 
 def add_arguments(parser):
@@ -55,8 +55,8 @@ class MegatronCheckpointLoaderLLM(MegatronCheckpointLoaderBase):
     def import_model_provider(self):
         """Return the correct model_provider function depending on GPT vs. BERT."""
         if self.args.model_type == 'GPT':
-            from model_provider import model_provider
-            from gpt_builders import gpt_builder
+            from megatron.training.models.gpt_builders import gpt_builder
+            from megatron.training.models.model_provider import model_provider
             self.model_provider = partial(model_provider, gpt_builder)
             return model_provider
         elif self.args.model_type == 'BERT':
